CourseBD/backend/requirements.txt
fastapi~=0.115.8
python-jose~=3.4.0
SQLAlchemy~=2.0.38
passlib~=1.7.4
pydantic~=2.10.6
uvicorn~=0.34.0
psycopg2~=2.9.10